requirements.txt
PyQt5==5.15.9
PyQt5-sip==12.13.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
urllib3==2.0.7
pyyaml==6.0.1
websockets==11.0.3
pyinstaller==5.13.2
xgboost==2.0.1
python-dateutil==2.8.2
pytz==2023.3

yfinance==0.2.48

jpholiday==0.1.10
